move modal callbacks to same file as composables
