--- app/src/main/java/com/piledrive/inventory/ui/callbacks/CreateTagCallbacks.kt
@@ -1,14 +1,0 @@
-package com.piledrive.inventory.ui.callbacks
-
-import com.piledrive.inventory.data.model.LocationSlug
-import com.piledrive.inventory.data.model.TagSlug
-
-interface CreateTagCallbacks {
-	//val onShowCreate: () -> Unit
-	val onAddTag: (slug: TagSlug) -> Unit
-}
-
-val stubCreateTagCallbacks = object : CreateTagCallbacks {
-	//override val onShowCreate: () -> Unit = {}
-	override val onAddTag: (slug: TagSlug) -> Unit = {}
-}
